chore: fix private filtering
